+++ yggdrasil_decision_forests/port/python/ydf/model/export_sklearn.py
@@ -61,50 +61,7 @@ ScikitLearnTree = TypeVar("ScikitLearnTree", bound=tree.BaseDecisionTree)
 
 
 def from_sklearn(sklearn_model: ScikitLearnModel) -> generic_model.GenericModel:
-  """Converts a tree-based scikit-learn model to a YDF model.
-
-  Usage example:
-
-  ```python
-  import ydf
-  from sklearn import datasets
-  from sklearn import tree
-
-  # Train a SKLearn model
-  X, y = datasets.make_classification()
-  skl_model = tree.DecisionTreeClassifier().fit(X, y)
-
-  # Convert the SKLearn model to a YDF model
-  ydf_model = ydf.from_sklearn(skl_model)
-
-  # Make predictions with the YDF model
-  ydf_predictions = ydf_model.predict(X)
-
-  # Analyse the YDF model
-  ydf_model.analyze(X)
-  ```
-
-  Currently supported models are:
-  *   sklearn.tree.DecisionTreeClassifier
-  *   sklearn.tree.DecisionTreeRegressor
-  *   sklearn.tree.ExtraTreeClassifier
-  *   sklearn.tree.ExtraTreeRegressor
-  *   sklearn.ensemble.RandomForestClassifier
-  *   sklearn.ensemble.RandomForestRegressor
-  *   sklearn.ensemble.ExtraTreesClassifier
-  *   sklearn.ensemble.ExtraTreesRegressor
-  *   sklearn.ensemble.GradientBoostingRegressor
-
-  Additionally, only single-label classification and scalar regression are
-  supported (e.g. multivariate regression models will not convert).
-
-  Args:
-    sklearn_model: the scikit-learn tree based model to be converted.
-
-  Returns:
-    a YDF Model that emulates the provided scikit-learn model.
-  """
-
+  """Converts a tree-based scikit-learn model to a YDF model."""
   if not hasattr(sklearn_model, "n_features_in_"):
     raise ValueError(
         "Scikit-Learn model must be fit to data before converting."